Add bytecode compiler
This adds a compiler, which writes the token tree data parsed from a
natscript code file to file and is able to later reconstruct that token
tree using that data. When later running the same file again, the
compiler will check if a compiled file corresponding to the code file is
present, and if yes, will load it from file instead of parsing the
source code. Loading the tokens from the compiled file is around 2.5x
faster than parsing them from the source code.